tests: fix ambuiguity when having two conftest files
