+++ tests/conftest.py
@@ -2,6 +2,8 @@ import json
 
 import pytest
 
+from mmore.privacy.agents.base import clear_llm_cache
+from mmore.privacy.agents.registry import tool_registry
 from mmore.type import MultimodalSample
 
 
@@ -27,3 +29,19 @@ def write_jsonl():
                 f.write(json.dumps(s.to_dict()) + "\n")
 
     return _write
+
+
+@pytest.fixture
+def isolate_llm_cache():
+    clear_llm_cache()
+    yield
+    clear_llm_cache()
+
+
+@pytest.fixture
+def isolated_tool_registry():
+    snapshot = dict(tool_registry)
+    tool_registry.clear()
+    yield
+    tool_registry.clear()
+    tool_registry.update(snapshot)
